refactor: centralize game types
