+++ src/components/game/CouncilPanel.tsx
@@ -1,34 +1,7 @@
 import React from 'react';
 import * as Dialog from '@radix-ui/react-dialog';
-import { GameResources } from './GameHUD';
+import { GameResources, CouncilProposal, ProposalDelta } from '@/types/game';
 import { getResourceIcon, getResourceColor } from './resourceUtils';
-
-export interface ProposalDelta {
-  grain?: number;
-  coin?: number;
-  mana?: number;
-  favor?: number;
-  unrest?: number;
-  threat?: number;
-}
-
-export interface CouncilProposal {
-  id: string;
-  title: string;
-  description: string;
-  type: 'economic' | 'military' | 'diplomatic' | 'mystical' | 'infrastructure';
-  cost: ProposalDelta;
-  benefit: ProposalDelta;
-  risk: number; // 0-100 percentage
-  duration: number; // cycles to complete
-  requirements?: string[];
-  canScry: boolean;
-  scryResult?: {
-    successChance: number;
-    hiddenEffects?: string[];
-  };
-  status?: 'pending' | 'accepted' | 'rejected' | 'applied';
-}
 
 export interface CouncilPanelProps {
   isOpen: boolean;
